Data Engineer Portfolio/Data Engin/requirements.txt


# Core Dependencies
apache-airflow==2.7.0
pandas==1.5.0
requests==2.28.0
python-dotenv==0.19.0
yfinance==0.1.70
psycopg2-binary==2.9.3
sqlalchemy==1.4.0

# Configuration & Utilities
pyyaml==6.0
python-dateutil==2.8.2
pytz==2022.7

# Email & Notifications
slack-sdk==3.20.0

# Monitoring
prometheus-client==0.16.0